+++ example/lib/db/seeders/seed_user_post.seeder.dart
@@ -24,6 +24,7 @@ class SeedUserPost extends SequelizeSeeding<CreatePost> {
           email: 'seed$index@example.com',
           firstName: 'Seed $index',
           lastName: 'User',
+          phoneNumber: SequelizeBigInt('${9000000000000 + index}'),
         ),
         postDetails: CreatePostDetails(
           likes: index,