+++ packages/sequelize_orm_generator/lib/src/generators/methods/_generate_find_all_method.dart
@@ -48,9 +48,11 @@ void _generateFindAllMethod(
   buffer.writeln('      sequelize: sequelizeInstance,');
   buffer.writeln('      model: sequelizeModel,');
   buffer.writeln('    ).then((results) =>');
-  buffer.writeln('      results.map((result) {');
+  buffer.writeln('      results.asMap().entries.map((entry) {');
+  buffer.writeln('        final rowIndex = entry.key;');
+  buffer.writeln('        final result = entry.value;');
   buffer.writeln(
-    '        final instance = $valuesClassName.fromJson(result.data);',
+    '        final instance = $valuesClassName.fromJson(result.data, operation: \'findAll\', rowIndex: rowIndex);',
   );
   buffer.writeln('        instance.originalQuery = query;');
   buffer.writeln('        instance.setPreviousDataValues(instance.toJson());');